Start adding more events to El
This obviously wont compile as some of the cases of `wrapElement` are not implemented but I'm having trouble building GHCJS.Dom on my system do to OSX / cairo difficulties so I cant compile. Also i thought you would have a better handle on whether or noe `mouseEventGetX` is the right function. At least its a start.