Race condition in websocket.
It is possible for reflex-dom to send a message into a socket that is still in a PENDING ReadyState, that is when it is building up a websocket connection.  When it does so, that causes a javascript error about trying to use an object that is not ready, breaking the whole page.

I'm not sure if this is technically a ghcjs-dom bug or a reflex-dom bug but it is certainly fixable here.