jsaddle-warp support for `run` and friends
The flag `use-warp` controls this. Because it is enabled by default,
the user would see a warp server in place of webkit/gtk frontend
when compiling the frontend in GHC.

Port number for the warp server is determined from the environment
variable `JSADDLE_WARP_PORT`, with a default value of 3003. The
port number will be printed to stdout by `run` so that user would
know where they can access the app.