Ask user for permissions.
From Android 6 on we have to ask for permissions at runtime. This commit
together with the PR for the reflex-platform asks the user for permission for
accessing camera and microphone as it is required from Android 6 onwards.

When the webview asks for permissions, we check whether we have the
corresponding Manifest permissions - if so, we simply grant the request. If not
we ask the user for permission. The implementation is also compatible with
Android 5 as we simply grant access in that case.

The split up into a PR for reflex-dom and reflex-platform is unfortunately
necessary because of the way we get the user's answer in Java: By means of an
overriden method of Activity.