Additional Hacking Attempted
I've attempted to fill in some of the blanks on this one and I think I
have a picture forming in my head of what we're trying to achieve. My
knowledge of Reflex with respect to the typeclasses is a bit thin so I'm
punching above my weight here. :)

I'll keep poking it. I think I can see how `withAsk` can be implemented
and `drawTemplate` should ideally fall out once we have the other pieces.

:)