reflex-dom-core: Eta expand some functions to appease the ghc 9 typechecker
This is a sad change, because hlint will complain about the unnecessary
eta expansion.

The problem is that ghc 9 will not unify the forall x. x with () in an
argument position. Although obviously a supplied value of the type
forall x. WidgetT x () can be used as a WidgetT () ().

eta expanding helps the type checker to agree with this.