Use https scheme and appassets.androidplatform.net instead of file:// for android.
This follows the usage recommended at
https://developer.android.com/reference/androidx/webkit/WebViewAssetLoader
to avoid an issue where assets referred to by file:// urls would receive
a permissions error and never be passed to shouldInterceptRequest in
sufficiently high API levels.

It may be appropriate to expose a method to use a different base URI,
which could simplify CORS for android apps in some use cases, and
additionally a full conversion to WebViewAssetLoader might be
reasonable.